AP_Compass: Field vector adjustment and online baseline correction
The intention patch is to solve the "incoherent compass error" in case people use several compass'.
In addition, this patch will improve the live plotting in GCS and makes sensor noise for compass instances more obvious as the scale of the read will be the same.
For this, I added per compass instance two LPF.
The first LPF will estimate the [x,y,z] baselines and try to correct the baseline offset according to the primary compass (which is not influenced).
The second LPF will try to scale the field vectors according to the primary compass rating after the baseline correction is performed.
These filters will not interfere with the heading estimate and the compass calibration,
because 1) the whole field vector will be scaled proportionally, and because 2) the primary compass will not be influenced.
The algorithm will not interfere with the other checks (alignment, ..) too.